requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.2
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-groq==0.1.0
langchain-huggingface==0.1.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
pydantic>=2.6.1
transformers>=4.37.2 